Stage 3/4: What's inside
In this stage your program should:

Ask for the number of the desired habitat using the following phrase: Please enter the number of the habitat you would like to view:

Use the input number as an index of your habitats to print its content.

End with the following phrase:

---
You've reached the end of the program. To check another habitat, please restart the watcher.