Stage 6/8: The value of life
Players start the game with eight lives. In other words, they can make a mistake only eight times.

When players input a valid letter, uncover it in the word, and carry on;
Print That letter doesn't appear in the word. and reduce the number of attempts if the word doesn't contain the letter, even if this particular letter has already been suggested before;
Print No improvements. and reduce the attempt counter when players input a letter that has been successfully uncovered before;
When players win, print the uncovered word, You guessed the word! , and the winning message. Each one should be on a new line.