Stage 7/8: Error
To complete this stage, follow the suggested order of the required checks:

Check whether players enter exactly one letter. If not, print Please, input a single letter.. Remember that when players input nothing or non-letter characters, it shouldn't count as a correct input either;
Make sure that the player entered a lowercase English letter. If not, the program should print Please, enter a lowercase letter from the English alphabet.;
If users enter the same letter twice (doesn't matter whether it appears in the word or not), then the program should output You've already guessed this letter.. Do not decrease the number of attempts in this case;
None of the three errors described above should reduce the number of remaining attempts!
When players win, print You guessed the word <word>!, where <word> should be substituted with the uncovered word, and the winning message. Each one should be on a new line.