Stage 6/6: The sandbox
The time goes on, and the competitors are developing their own coffee machines to attract coffee lovers with special types of coffee or interesting features. It's time to enhance your coffee machine to compete with them!

In this stage, you are free to add anything you want to your coffee machine!

Objective
This stage will not be tested. Try to be creative and add something new to your coffee machine. It can be a new type of coffee, different milk flavors, or different volumes of disposable cups.

Leave a comment with the link to the full source code of your coffee machine and tell us what you have implemented!

Good luck!